fix: render page-level props on every `.md` projection
