+++ CHANGELOG.md
@@ -3,7 +3,28 @@
 All notable changes to outl are documented here.
 Format inspired by [Keep a Changelog](https://keepachangelog.com/en/1.1.0/); the project uses [Semantic Versioning](https://semver.org/).
 
-## [0.6.0] — 2026-06-05
+## [Unreleased]
+
+### Added
+
+- **`@` mention autocomplete** — typing `@` in any block opens a person picker filtered to pages with the `type:: person` page-level property, fuzzy-matched against the typed name. Accepting a candidate inserts `[[@name]]` — a regular wikilink whose `@` is the link affordance only (page identity stays clean, slug has no `@`). Composite names like `@Thiago Avelino` work because the autocomplete query allows spaces. A "create new" candidate appears when the typed query doesn't match any existing person; accepting it materialises a fresh `pages/<slug>.md` with `type:: person` already set, so the next mention of the same name surfaces it without manual property editing. Wired identically in `outl-tui`, `outl-desktop`, and `outl-mobile`; the shared `@outl/shared/autocomplete` library owns the trigger detection + create-new helper for the GUI clients.
+- **`type::` page-level property** — surfaced on `outl_md::WorkspaceIndex::PageEntry.page_type` and `outl_actions::PageMeta.page_type`. New filter `WorkspaceIndex::pages_by_type(t)` and consumer `outl_actions::page::search_persons(ws, query)` rank `type:: person` pages for the `@` mention popup. `type::` is just one of many user-facing page properties — `title::`, `icon::`, `pinned::`, `role::`, anything custom — and all of them now reach the workspace tree (see "Fixed" below).
+- **`ref-projection-failed` Tauri event** — emitted by the desktop and mobile clients when `open_ref` resolved a target (page is in the op log) but writing the resulting `.md` + sidecar failed. Frontend can listen via `onRefProjectionFailed` (desktop) and surface a toast so the user knows the link they just inserted isn't visible to peers yet. The op log retry on the next save / orphan scan is still automatic.
+
+### Fixed
+
+- **Page-level properties now reach the workspace tree.** The reconcile pipeline previously emitted `Op::SetProp` only for **block-nested** properties; anything written at the top of a `.md` (above the first `-` bullet) — `title::`, `icon::`, `pinned::`, `type::`, custom keys — lived only on disk. The workspace CRDT never learned about them, so any consumer reading via `workspace.tree().property(page_id, …)` (desktop sidebar, mobile picker, `outl_actions::search_persons`) silently disagreed with the rendered markdown. The TUI hid the bug because its `WorkspaceIndex`-backed surfaces parse `.md` straight from disk. Cross-client divergence on every workspace populated outside the in-app picker — fixtures, vim users, Logseq/Roam imports, peers via iCloud — was the result. Fix: `outl_md::diff::diff_to_ops_with_page_props` emits `Op::SetProp` on the page root for every page-level property in the parsed AST; `outl_md::reconcile::reconcile_md` calls it on every reconcile pass.
+- **Page root now materialises in the tree.** Pages authored externally never received an `Op::Create`, only the blocks **under** them did. Combined with the CRDT contract that `Op::Move` on a node without a preceding `Op::Create` is a no-op (an intentional design for peer-sync ordering), this left the page node as a ghost — present as `parent` of its blocks but absent from `children_of(NodeId::root())`. `list_all_pages`, `search_persons`, and the sidebar all skipped externally-authored pages silently. Fix: `outl_md::reconcile::ensure_page_root_in_tree` emits `Op::Create` when the page node is absent from `self.nodes`, or `Op::Move` when it exists at the wrong parent, plus `Op::SetProp` for `page-slug` / `page-kind`. Idempotent — 0 ops emitted on pages that are already materialised.
+- **`open_ref` regenerates `.md` after creating a page.** Both desktop and mobile previously left newly-created pages on the op log only; the `pages/<slug>.md` projection never landed on disk until something else triggered `apply_page_md_with_sidecar` on that page. `WorkspaceIndex` (which parses `.md` from disk) disagreed with the tree CRDT silently, and a peer pulling the workspace via iCloud would never see the page at all. Fix: both clients now call `apply_page_md_with_sidecar` immediately after the `open_or_create_by_ref` mutation; failures emit `ref-projection-failed`.
+- **`open_or_create_by_ref` no longer drops the `@` arm via slug normalisation.** `slugify("@avelino")` strips the `@` and returns `"avelino"`. The generic `find_by_slug(slugify(target))` branch used to run **before** the `@` arm, so a pre-existing `pages/avelino.md` (created before this feature, or by an external editor without `type:: person`) resolved via the generic path and returned early — never marking the page as a person. Fix: the `@` arm runs first and idempotently sets `type:: person` on every resolution, even when the page already existed.
+- **`reconcile_md` reads the sidecar once.** The short-circuit check used to re-read the sidecar file separately from the diff inputs — racy if another process rewrote the sidecar between the two reads. Fix: single read, both consumers share the result.
+- **Background-thread reconcile after open.** Opening a workspace used to block on a synchronous `reconcile_md` pass across every legacy page. With many pages, the first paint waited tens of seconds. Fix: `outl-desktop::workspace_open::spawn_background_reconcile` runs the orphan reconcile on a separate thread, locks the workspace per page (released between iterations), and emits `workspace-reconciled` on completion. Today's journal opens immediately; legacy pages materialise behind the scenes.
+
+### Migration
+
+- **`pipeline_version` in the sidecar drives forward-compatible re-reconciles.** The first boot on an upgraded binary scans every `pages/<slug>.outl` and re-runs `reconcile_md` on any sidecar whose `pipeline_version` is below the binary's `CURRENT_PIPELINE_VERSION`. Idempotent: the pipeline emits the same `Op::Create` / `Op::SetProp` ops that would have been emitted on first ingestion of the `.md`, the CRDT deduplicates by LWW, and the sidecar persists the bumped version. Two clients opening the same legacy workspace at the same time will each run their own reconcile (each actor owns its `ops-<actor>.jsonl`); the log inflates by roughly 2× the page-root ops once per legacy page per device — aceitável for a one-shot migration, the CRDT converges deterministically. Subsequent boots skip the page via the `last_synced_hash` + `pipeline_version` short-circuit.
+
+
 
 **Desktop client ships.**
